Without the bin feature and no integration tests, building a separate rlib provides little benefits besides increased build durations.
